Check if a OneDrive exists
This script will take a csv and compare it against a M365 tenant and check to see if the user has a OneDrive provisioned. This can be helpful if you are doing migrations from tenant to tenant or if you just want to see who still has not used OneDrive yet. This uses PowerShell 5.1 and an older version of PnP.PowerShell.